+++ src/types-internal/editor-modules.d.ts
@@ -5,7 +5,6 @@ import InlineToolbar from '../components/modules/toolbar/inline';
 import Toolbox from '../components/modules/toolbar/toolbox';
 import BlockSettings from '../components/modules/toolbar/blockSettings';
 import Paste from '../components/modules/paste';
-import Notifier from '../components/modules/notifier';
 import DragNDrop from '../components/modules/dragNDrop';
 import ModificationsObserver from '../components/modules/modificationsObserver';
 import Renderer from '../components/modules/renderer';
@@ -52,7 +51,6 @@ export interface EditorModules {
   API: API;
   Caret: Caret;
   Saver: Saver;
-  Notifier: Notifier;
   BlockManager: BlockManager;
   BlocksAPI: BlocksAPI;
   CaretAPI: CaretAPI;